requirements.txt
openai-whisper
pydub
ffmpeg-python
click
rich
pathlib
